break up commands
