+++ packages/xreal_one_ctl/src/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod enable_camera;
+pub mod flash_dsp;
+pub mod get_config;
+pub mod info;
+pub mod recover_mcu;
+pub mod update;
